tests: Split Brownian/Langevin/DPD tests
Move checks with long runtime into separate files.
