+++ testsuite/python/CMakeLists.txt
@@ -130,7 +130,9 @@ python_test(FILE ek_eof_one_species_y.py MAX_NUM_PROC 1 LABELS gpu)
 python_test(FILE ek_eof_one_species_z.py MAX_NUM_PROC 1 LABELS gpu)
 python_test(FILE exclusions.py MAX_NUM_PROC 2)
 python_test(FILE langevin_thermostat.py MAX_NUM_PROC 1)
+python_test(FILE langevin_thermostat_stats.py MAX_NUM_PROC 1)
 python_test(FILE brownian_dynamics.py MAX_NUM_PROC 1)
+python_test(FILE brownian_dynamics_stats.py MAX_NUM_PROC 1)
 python_test(FILE nsquare.py MAX_NUM_PROC 4)
 python_test(FILE virtual_sites_relative.py MAX_NUM_PROC 2)
 python_test(FILE virtual_sites_tracers.py MAX_NUM_PROC 2)
@@ -142,6 +144,7 @@ python_test(FILE dipolar_mdlc_p3m_scafacos_p2nfft.py MAX_NUM_PROC 1 LABELS long)
 python_test(FILE lb.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE force_cap.py MAX_NUM_PROC 2)
 python_test(FILE dpd.py MAX_NUM_PROC 4)
+python_test(FILE dpd_stats.py MAX_NUM_PROC 4)
 python_test(FILE hat.py MAX_NUM_PROC 4)
 python_test(FILE analyze_energy.py MAX_NUM_PROC 2)
 python_test(FILE analyze_mass_related.py MAX_NUM_PROC 4)
